refactor: change routes config
