--- src/pages/index.tsx
@@ -1,2 +1,0 @@
-export { RegistrationEnd } from "./RegistrationEnd";
-export { Home } from "./Home";
